chore updating utils file
