+++ src/library/utils.ts
@@ -32,7 +32,19 @@ const OPERATOR_ALIASES = {
   $or: Op.or
 };
 
+export interface Query {
+  limit?: number;
+  offset?: number;
+  order?: any[];
+  skip?: number;
+  include?: any[];
+  where?: any;
+  attributes?: any;
+}
 
+export function getModelFromList(db: Sequelize, modelName) {
+  return db.models[modelName];
+}
 
 export function sanitizeAttributes(attributes: any): any {
   // If `attributes` parameter is a string, try to interpret it as JSON
@@ -221,4 +233,98 @@ export function parseAttributes(req: Request): any {
 
   // Return final `attributes`.
   return attributes;
+}
+
+export function parseInclude(
+  req: Request,
+  model: ModelCtor<any>,
+  db: Sequelize
+): Array<any> {
+  try {
+    let include: Array<any> = [];
+    const populate: any = req.query.include || req.query.populate;
+
+    if (_.isString(populate)) {
+      include = JSON.parse(populate);
+    }
+
+    if (!Array.isArray(include)) {
+      throw ControllerErrors.BAD_REQUEST;
+    }
+
+    const tryWithFilter = (m: string, model: ModelCtor<any>) => {
+      if (!m.length) {
+        throw ControllerErrors.BAD_REQUEST;
+      }
+      /*
+       Two options here:
+       1. We have a Model name (like User) or a Model name with filter (like User.filter)
+       2. We have the name of the property for the association (like 'user' or 'owner')
+
+       1 always starts with uppercase, 2 with lowercase
+      */
+
+      const start = m[0];
+
+      if (start === start.toLowerCase()) {
+        // 2. name of the property
+        // Get association data from model
+        const association = model.associations[m];
+        if (association == null) {
+          throw ControllerErrors.BAD_REQUEST;
+        }
+        const targetModel = association.target;
+        const as = association.as;
+        return { model: targetModel, as, required: false };
+      }
+
+      // 1. We have the Model name
+      if (m.includes(".")) {
+        const splt = m.split("."),
+          modelName = splt[0],
+          filterName = splt[1];
+
+        const model = getModelFromList(db, modelName);
+
+        //return {model: model, where: where, required: false};
+        if (model["filter"] != null) {
+          return model["filter"](filterName);
+        }
+      }
+
+      return { model: getModelFromList(db, m), required: false };
+    };
+
+    const parseIncludeRecursive = (item, model: ModelCtor<any>) => {
+      if (_.isString(item)) {
+        // Simple include
+        return tryWithFilter(item, model);
+      } else {
+        // Include with nested includes
+        const modelName: string = Object.keys(item)[0];
+        const content = item[modelName];
+
+        if (!Array.isArray(content)) {
+          throw ControllerErrors.BAD_REQUEST;
+        }
+
+        const result: any = tryWithFilter(modelName, model);
+        result.include = content.map(i =>
+          parseIncludeRecursive(i, result.model)
+        );
+
+        return result;
+      }
+    };
+
+    let preparedInclude = include.map(item =>
+      parseIncludeRecursive(item, model)
+    );
+    // Merge with req.session.include (Useful for enforcing policies)
+    preparedInclude = _.union(preparedInclude, req.session?.include || {});
+    return preparedInclude;
+  } catch (err) {
+    console.log(err)
+    throw ControllerErrors.BAD_REQUEST;
+  }
 }
